frontend/webserver/requirements.txt
Flask>=2.0.3
boto3>=1.21.33
bootstrap-flask>=2.1.0
kafka-python>=2.0.2
awscli>=1.20.0